Add color scheme options to visualization
This commit defines two color schemes (DAYLIGHT_MODE and DEFAULT_MODE) for the visualization in the main.py file. The selected scheme is applied to all visual elements, such as text and matrices. This makes it easy to switch between color schemes, simply by changing the value of the MODE variable.